move code to dedicated conventional commits file
