+++ internal/nsv/semver_test.go
@@ -21,221 +21,3 @@ SOFTWARE.
 */
 
 package nsv_test
-
-import (
-	"testing"
-
-	git "github.com/purpleclay/gitz"
-	"github.com/purpleclay/nsv/internal/nsv"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestDetectIncrement(t *testing.T) {
-	tests := []struct {
-		name     string
-		commit   string
-		expected nsv.Increment
-		// TODO: check position
-	}{
-		{
-			name:     "NoIncrement",
-			commit:   "chore(deps): bump action from 1.0.1 to 1.1.0",
-			expected: nsv.NoIncrement,
-		},
-		{
-			name:     "FixPatchIncrement",
-			commit:   "fix: incorrect cache retrieval based on key",
-			expected: nsv.PatchIncrement,
-		},
-		{
-			name:     "FixScopedPatchIncrement",
-			commit:   "fix(ui): paging issue within the search table",
-			expected: nsv.PatchIncrement,
-		},
-		{
-			name:     "FeatMinorIncrement",
-			commit:   "feat: add new grep feature to cli",
-			expected: nsv.MinorIncrement,
-		},
-		{
-			name:     "FeatScopedMinorIncrement",
-			commit:   "feat(search): add fuzzy finding to predictive search",
-			expected: nsv.MinorIncrement,
-		},
-		{
-			name:     "BangMajorIncrement",
-			commit:   "refactor!: rename fields within cli context",
-			expected: nsv.MajorIncrement,
-		},
-		{
-			name:     "BangScopedMajorIncrement",
-			commit:   "fix(api)!: rewrite of existing API",
-			expected: nsv.MajorIncrement,
-		},
-		{
-			name: "BreakingFooterMajorIncrement",
-			commit: `perf: rewrite sorting algorithm and public interface
-BREAKING CHANGE: this is a breaking change`,
-			expected: nsv.MajorIncrement,
-		},
-		{
-			name: "HyphenatedBreakingFooterMajorIncrement",
-			commit: `fix(ui): incorrect sizing of components during window resize
-BREAKING-CHANGE: this is a breaking change`,
-			expected: nsv.MajorIncrement,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			inc, _ := nsv.DetectIncrement([]git.LogEntry{
-				{
-					Message: tt.commit,
-				},
-			})
-			require.Equal(t, tt.expected, inc, "failed to match increment")
-		})
-	}
-}
-
-func TestDetectIncrementFindsHighestIncrement(t *testing.T) {
-	log := []git.LogEntry{
-		{Message: "ci: support dependency scanning in workflows"},
-		{Message: "feat: support dependency injection of database tier"},
-		{Message: "fix: tui not redrawing correctly during terminal window resize"},
-		{Message: "feat!: depecrate existing api and switch to GraphQL"},
-		{
-			Message: `refactor(cli): remove use of existing flags and rename
-BREAKING CHANGE: this is a breaking change`,
-		},
-		{Message: "docs: build documentation using material for mkdocs"},
-	}
-	inc, pos := nsv.DetectIncrement(log)
-
-	assert.Equal(t, nsv.MajorIncrement, inc)
-	assert.Equal(t, 3, pos)
-}
-
-func TestDetectIncrementStrictness(t *testing.T) {
-	tests := []struct {
-		name     string
-		commit   string
-		expected nsv.Increment
-	}{
-		{
-			name:     "NoSpaceAfterTypeColon",
-			commit:   "feat:add support for multiple search terms",
-			expected: nsv.NoIncrement,
-		},
-		{
-			name: "NoSpaceAfterBreakingColon",
-			commit: `refactor: add caching layer in front of database
-BREAKING CHANGE:this is a breaking change`,
-			expected: nsv.NoIncrement,
-		},
-		{
-			name: "NoSpaceAfterHyphenatedBreakingColon",
-			commit: `perf: change to write through caching
-BREAKING-CHANGE:this is a breaking change`,
-			expected: nsv.NoIncrement,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			inc, _ := nsv.DetectIncrement([]git.LogEntry{
-				{
-					Message: tt.commit,
-				},
-			})
-			require.Equal(t, tt.expected, inc, "failed to match increment")
-		})
-	}
-}
-
-func TestDetectIncrementCaseInsensitiveLabel(t *testing.T) {
-	tests := []struct {
-		name     string
-		commit   string
-		expected nsv.Increment
-	}{
-		{
-			name:     "ForFixType",
-			commit:   "FiX(ui): incorrect styling of elements on page",
-			expected: nsv.PatchIncrement,
-		},
-		{
-			name:     "ForFeatType",
-			commit:   "FeAt: add tui support using bubbletea",
-			expected: nsv.MinorIncrement,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			inc, _ := nsv.DetectIncrement([]git.LogEntry{
-				{
-					Message: tt.commit,
-				},
-			})
-			require.Equal(t, tt.expected, inc, "failed to match increment")
-		})
-	}
-}
-
-func TestDetectIncrementCaseSensitiveBreaking(t *testing.T) {
-	tests := []struct {
-		name     string
-		commit   string
-		expected nsv.Increment
-	}{
-		{
-			name: "MixedCase",
-			commit: `fix: incorrectly closing handles to database
-Breaking Change: this is a breaking change`,
-			expected: nsv.PatchIncrement,
-		},
-		{
-			name: "MixedCaseHyphenated",
-			commit: `feat: switched from mongodb to surrealdb
-Breaking-Change: this is a breaking change`,
-			expected: nsv.MinorIncrement,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			inc, _ := nsv.DetectIncrement([]git.LogEntry{
-				{
-					Message: tt.commit,
-				},
-			})
-			require.Equal(t, tt.expected, inc, "failed to match increment")
-		})
-	}
-}
-
-// globals are used to prevent any compiler optimizations
-var gInc nsv.Increment
-var gPos int
-
-func BenchmarkDetectIncrement(b *testing.B) {
-	log := []git.LogEntry{
-		{Message: "docs:"},
-		{Message: "ci:"},
-		{Message: "feat(parser):"},
-		{Message: "chore:"},
-		{Message: "feat(ui):"},
-		{Message: "fix:"},
-		{Message: "fix(ui):"},
-		{Message: "feat:"},
-		{Message: "ci:"},
-		{Message: "refactor:"},
-	}
-	b.ResetTimer()
-
-	var inc nsv.Increment
-	var pos int
-	for n := 0; n < b.N; n++ {
-		inc, pos = nsv.DetectIncrement(log)
-	}
-	gInc = inc
-	gPos = pos
-}
